[DOP-21799] Refactor celery initialization
