+++ syncmaster/scheduler/celery.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
+# SPDX-License-Identifier: Apache-2.0
+from syncmaster.scheduler.settings import SchedulerAppSettings
+from syncmaster.worker import celery_factory
+
+celery = celery_factory(SchedulerAppSettings())
