+++ tests/test_integration/celery_test.py
@@ -1,3 +1,3 @@
-from syncmaster.worker import celery
+from syncmaster.scheduler.celery import celery
 
 celery.conf.update(imports=list(celery.conf.imports) + ["tests.test_integration.test_scheduler.test_task"])
